signa-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
passlib==1.7.4
argon2-cffi==25.1.0
bcrypt==4.3.0
PyJWT==2.10.1